Change the tint of green used for satisfied and add title attributes
On some displays the darker green in combination with the blue text around it will look blue-ish, rather than the more telling green.
The lime color keyword is widely supported and a clearer green indicator in this case.

This doesn't take away from the fact that indicators based on color with no other indicator *will* give issues for people with color blindness, no matter what.
So to improve usability, I've added a title attribute to the links in the list items indicating the status.